Add JWT authentication

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,19 @@
-import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 import { formatDistanceToNow } from 'date-fns';
+import { useAuthContext } from '../hooks/useAuthContext';
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 
 export const WorkoutDetails = ({ workout }) => {
 	const { dispatch } = useWorkoutsContext();
+	const { user } = useAuthContext();
 
 	const handleClick = async () => {
+		if (!user) return;
+
 		const response = await fetch(`/api/workouts/${workout._id}`, {
 			method: 'DELETE',
+			headers: {
+				Authorization: `Bearer ${user.token}`,
+			},
 		});
 
 		const json = await response.json();
